vector-store: move full-scan range scheduling into full_scan module
Statements::initial_scan mixed deriving the token ranges and opening the
range queries (which need a ScyllaDB session) with scheduling the ranges
under a concurrency limit and tracking when the pipeline acknowledged
their rows. The scheduling part is where the full scan spends its time,
but it could not be exercised without a live cluster: the integration
harness replaces the whole db actor with a mock.

Move the scheduling loop verbatim into db_index/full_scan.rs as a
generic scan_ranges() that takes the ranges, the concurrency limit, an
"open range" closure and a range-length closure, so that the following
patches can test it with fake streams. initial_scan() becomes a thin
adapter around it. No behaviour change.

Refs: VECTOR-862
